Fix template prototype
